refactor: split workflows controller into crud/blueprints/validation package per ADR-0008 (#2253)
