+++ src/synthorg/api/controllers/workflows/__init__.py
@@ -1,0 +1,8 @@
+"""Workflow definition controllers.
+
+Split per ADR-0008 into per-concern controller sub-modules sharing the
+``/workflows`` path: :mod:`crud` (listing + mutations), :mod:`blueprints`
+(blueprint listing + instantiation), and :mod:`validation` (draft/stored
+validation + YAML export). The shared per-request service factory lives
+in :mod:`_shared`. Import controllers directly from their sub-module.
+"""
